+++ src/main/java/com/codeborne/selenide/proxy/InetAddressResolver.java
@@ -1,0 +1,15 @@
+package com.codeborne.selenide.proxy;
+
+import java.net.InetAddress;
+import java.net.UnknownHostException;
+
+public class InetAddressResolver {
+  InetAddress getInetAddressByName(String hostname) {
+    try {
+      return InetAddress.getByName(hostname);
+    }
+    catch (UnknownHostException e) {
+      throw new IllegalArgumentException(e);
+    }
+  }
+}