+++ src/test/java/com/codeborne/selenide/proxy/SelenideProxyServerTest.java
@@ -1,11 +1,14 @@
 package com.codeborne.selenide.proxy;
 
-import java.net.InetSocketAddress;
-
+import com.codeborne.selenide.Configuration;
 import net.lightbody.bmp.BrowserMobProxyServer;
 import org.assertj.core.api.WithAssertions;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.openqa.selenium.Proxy;
+
+import java.net.InetSocketAddress;
 
 import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.mock;
@@ -14,29 +17,46 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 class SelenideProxyServerTest implements WithAssertions {
+  private BrowserMobProxyServer bmp = mock(BrowserMobProxyServer.class);
+  private SelenideProxyServer proxyServer = new SelenideProxyServer(null, new InetAddressResolverStub(), bmp);
+
+  @BeforeEach
+  @AfterEach
+  void resetProxySettings() {
+    Configuration.proxyPort = 0;
+    Configuration.proxyHost = "";
+  }
+
   @Test
   void canInterceptResponses() {
-    BrowserMobProxyServer bmp = mock(BrowserMobProxyServer.class);
-    when(bmp.getPort()).thenReturn(8888);
-
-    SelenideProxyServer proxyServer = new SelenideProxyServer(null);
-    proxyServer.proxy = bmp;
     proxyServer.start();
 
-    try {
-      verify(bmp).setTrustAllServers(true);
-      verify(bmp, never()).setChainedProxy(any(InetSocketAddress.class));
-      verify(bmp).start();
-      assertThat(proxyServer.createSeleniumProxy().getHttpProxy())
-        .endsWith(":8888");
-    } finally {
-      proxyServer.shutdown();
-    }
-    verify(bmp).abort();
+    verify(bmp).setTrustAllServers(true);
+    verify(bmp, never()).setChainedProxy(any(InetSocketAddress.class));
+    verify(bmp).start(0);
 
     FileDownloadFilter filter = proxyServer.responseFilter("download");
-    assertThat(filter.getDownloadedFiles().size())
-      .isEqualTo(0);
+    assertThat(filter.getDownloadedFiles()).hasSize(0);
+  }
+
+  @Test
+  void canShutdownProxyServer() {
+    proxyServer.shutdown();
+    verify(bmp).abort();
+  }
+
+  @Test
+  void createSeleniumProxy() {
+    when(bmp.getPort()).thenReturn(8888);
+
+    assertThat(proxyServer.createSeleniumProxy().getHttpProxy()).endsWith(":8888");
+  }
+
+  @Test
+  void createSeleniumProxy_withConfiguredHostname() {
+    Configuration.proxyHost = "my.megahost";
+    when(bmp.getPort()).thenReturn(9999);
+    assertThat(proxyServer.createSeleniumProxy().getHttpProxy()).isEqualTo("my.megahost:9999");
   }
 
   @Test
@@ -46,9 +66,16 @@ class SelenideProxyServerTest implements WithAssertions {
 
     InetSocketAddress proxyAddress = SelenideProxyServer.getProxyAddress(proxy);
 
-    assertThat(proxyAddress.getHostName())
-      .isEqualTo("111.22.3.4444");
-    assertThat(proxyAddress.getPort())
-      .isEqualTo(8080);
+    assertThat(proxyAddress.getHostName()).isEqualTo("111.22.3.4444");
+    assertThat(proxyAddress.getPort()).isEqualTo(8080);
+  }
+
+  @Test
+  void canStartProxyServerOnConfiguredPort() {
+    Configuration.proxyPort = 666;
+
+    proxyServer.start();
+
+    verify(bmp).start(666);
   }
 }